requirements.txt
Flask==3.0.0
flask-cors==4.0.0
yfinance>=0.2.33
pandas>=2.1.4
numpy>=1.26.2
scikit-learn>=1.3.0
newsapi-python>=0.2.7
python-dotenv>=1.0.0
google-generativeai>=0.8.0
langchain
langchain-google-genai
langchain-community